feat: update sketchybar and yabai config spaces
This is the beginning of a long clean up of the Yabai and Sketchybar
configuration. The end goal is to turn Yabai into i3. This push enables
space icons to be automatically created when a given app is present.
But, it does not automatically create the space and put an app on it,
yet.