feat: add auto resizing to sketchybar
Added a helper function to sketchybar which finds the current menu bar
height. This is run on launch. Added a yabai signal to detect when the
resolution is changed and tell sketchybar to restart when this happens.

This updates the bar height to be the height of the notch regardless of
the resolution.