fix: disable childframe for company
I had issues with this due to my Yabai configuration and the "fixes"
that I apply. Whenever I use Doom's big-font mode, it resizes the frame
for some reason. This causes it to be unmanaged and unfocused in Yabai.
I rebalance the space whenever this happens. Unfortunately, the
childframe causes this to active as well, so I can live without it.