perf: use Emacs daemon launcher and add robustness
Removed operating system detection and switched to using Emacs' daemon
and emacsclient commands themselves to manage the daemon on each OS.
Then, switch to a singular emacsclient command to both launch the daemon
and connect to it. This removes any sleeping or checking for the daemon
to be ready to connect to resulting in no wasted time waiting for and
connecting to the server.