refactor: update i3 config and emacs mode format
This refactor was done to standardize the layout of the documentation
for my i3 window manager in order to sync it with my Yabai window
manager config. I've embraced Org's tangling and noweb referencing to be
able to have the emacs mode keybind pass through in an organized manner.
I've also removed apps and workspaces I no longer use.