fix: test method to get all org files to show up
I'm trying to ensure that all the org files in .config show up when I
click on the org language button on GitHub to initiate a search. I may
end up wanting to include/exclude certain files which contain no
valuable code and truly are just documentation in the future but for now
I would like everything to just show up.