feat: implement i3 wm scratchpad in yabai
This works by creating a space dedicated to acting as the "scratchpad"
and moving windows to and from it. To increase the speed of moving these
windows the window id is cached in its own file to avoid having to query
yabai and using jq to lookup values in the array.

This is all done using signals from yabai. Upon window creation we store
the id for the window. So far, this method works great. The only issues
spotted so far are inconsistencies with some apps not dying upon window
deletion leading to a somewhat "stuck" state, but there's an easy fix
for that.