feat(org-roam): add automations to file tags
When linking to a person, @UserName tag is automatically added. Whenever
actionable todo keywords (non-typed) are added to an org-roam node the
file is automatically tagged as "agenda" and included in the agenda
files. When a file is tagged "person", the @UserName tag is
automatically added to the file too.