feat(org): add org filetag sorting and filename syncing
This uses alphabetical sorting for filetags and Denote's file name
keywords which are triggered upon saving after modifying the tags.

Someday, I may want to change this behavior to not just trigger on
filetag changes but other front matter. Additionally, on-save may not be
the best hook to use for filetags, instead tag adding could be better.