refactor(i3): create better document generation
Prior to this version, the i3 config generation was a bit of a hack.
This version although still a bit complex, should be easier to maintain
and understand, as well as be cleaner for the user to read either the
config file itself (generated comments) or the Org file for its
categorical overview.

This update also includes a number of small changes (I know I should
make more commits) such as update settings names, changing of windowing
styles for Discord and Spotify, and removal/hiding of unused code.